Extract Win32 DebugOutputWindow reporter to doctest/parts/reporters/debug_output_window.h
